backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
python-dotenv==1.0.1
PyMuPDF==1.24.9
pdfplumber==0.11.0
camelot-py==0.11.0
opencv-python==4.10.0.84
pytesseract==0.3.13
pandas==2.2.2
numpy==1.26.4
django-cors-headers==4.9.0
